fix auth

+++ managers/entities/student/student.manager.js
@@ -1,3 +1,4 @@
+const ROLES = require('../../../helpers/roles');
 /**
  * The StudentManager class is responsible for managing student records in the database.
  * @class
@@ -22,29 +23,30 @@ class StudentManager {
      * @param {string} params.firstName - The first name of the student.
      * @param {string} params.lastName - The last name of the student.
      * @param {string} params.classroomId - The ID of the classroom the student belongs to.
+     * @param {string} params.schoolId - The ID of the school the student belongs to.
      * @param {string} params.username - The username for the student's account.
      * @param {string} params.password - The password for the student's account.
      * @param {Object} params.__longToken - The token containing user role information.
      * @returns {Promise<Object>} The created student object or an error object.
      * @throws Will throw an error if there is an issue during the creation process.
      */
-    async create({firstName, lastName, classroomId,username,password, __longToken}){
+    async create({firstName, lastName, classroomId,schoolId,username,password, __longToken}){
         try {
             const userRole = __longToken?.role;
             const isAuthorized = this.managers.authorization.isAuthorized({userRole, action: 'create', resource: 'student'});
             if(!isAuthorized) return {error: 'Unauthorized', status: 401, ok: false};
-           
-            const student = {firstName, lastName, classroomId,username,password};
+            if(userRole===ROLES.SCHOOL_ADMIN && schoolId!==__longToken.schoolId) return {error: 'Unauthorized', status: 401, ok: false};
+
+            const student = {firstName, lastName, classroomId,username,password,schoolId};
             let result = await this.validators.student.createStudent(student);
             if(result) return result;
 
             // Check if classroom exists
             const classroom = await this.mongomodels.classroom.findById(classroomId);
             if(!classroom) return {error: 'classroom not found', status: 404, ok: false};
-
-            // Creation Logic
-            // create user
-            // check if username exists
+            const school = await this.mongomodels.school.findById(schoolId);
+            if(!school) return {error: 'school not found', status: 404, ok: false};
+           
             let userExists = await this.mongomodels.user.findOne({username});
             if(userExists) return {error: 'User already exists', status: 409,ok: false};
             let user = {username, password, role: 'student'};
@@ -54,7 +56,8 @@ class StudentManager {
                 firstName, 
                 lastName, 
                 classroomId, 
-                userId: createdUser._id
+                userId: createdUser._id,
+                schoolId
                 });
 
             // Response
@@ -80,10 +83,11 @@ class StudentManager {
      * @param {string} params.firstName - The first name of the student.
      * @param {string} params.lastName - The last name of the student.
      * @param {string} params.classroomId - The ID of the classroom the student belongs to.
+     * @param {string} params.schoolId - The ID of the school the student belongs to.
+
      * @param {string} params.id - The ID of the student.
      * @param {string} [params.username] - The username of the student.
      * @param {Object} params.__longToken - The token containing user role information.
-     * @param {string} params.__longToken.role - The role of the user making the request.
      *
      * @returns {Promise<Object>} The result of the update operation.
      * @returns {Object} [returns.error] - The error message if the update fails.
@@ -97,13 +101,15 @@ class StudentManager {
      *
      * @throws {Error} If an error occurs during the update process.
      */
-    async update({firstName, lastName, classroomId, id,username, __longToken}){
+    async update({firstName, lastName, classroomId, schoolId,id,username, __longToken}){
         try {
             const userRole = __longToken?.role;
             const isAuthorized = this.managers.authorization.isAuthorized({userRole, action: 'update', resource: 'student'});
             if(!isAuthorized) return {error: 'Unauthorized', status: 401, ok: false};
+            if(userRole===ROLES.SCHOOL_ADMIN && schoolId!==__longToken.schoolId) return {error: 'Unauthorized', status: 401, ok: false};
+
           
-            const student = {firstName, lastName, classroomId, id};
+            const student = {firstName, lastName, classroomId, id,schoolId,username};
             let result = await this.validators.student.updateStudent(student);
             if(result) return result;
 
@@ -113,7 +119,8 @@ class StudentManager {
             // Check if classroom exists
             const classroom = await this.mongomodels.classroom.findById(classroomId);
             if(!classroom) return {error: 'classroom not found', status: 404, ok: false};
-
+            const school = await this.mongomodels.school.findById(schoolId);
+            if(!school) return {error: 'school not found', status: 404, ok: false};
             if(username){
                 // Check if username exists
                 let userExists = await this.mongomodels.user.findOne({username, _id: {$ne: studentExists.userId}});
@@ -168,6 +175,7 @@ class StudentManager {
 
             const studentExists = await this.mongomodels.student.findById(id);
             if(!studentExists) return {error: 'student not found', status: 404, ok: false};
+            if(userRole===ROLES.SCHOOL_ADMIN && studentExists.schoolId!==__longToken.schoolId) return {error: 'Unauthorized', status: 401, ok: false};
 
             // Deletion Logic
             let deletedStudent = await this.mongomodels.student.findByIdAndDelete(id);
@@ -204,6 +212,7 @@ class StudentManager {
 
             const studentExists = await this.mongomodels.student.findById(id);
             if(!studentExists) return {error: 'student not found', status: 404, ok: false};
+            if(userRole===ROLES.SCHOOL_ADMIN && studentExists.schoolId!==__longToken.schoolId) return {error: 'Unauthorized', status: 401, ok: false};
 
             return {
                 student: studentExists
@@ -227,6 +236,13 @@ class StudentManager {
             const userRole = __longToken?.role;
             const isAuthorized = this.managers.authorization.isAuthorized({userRole, action: 'create', resource: 'student'});
             if(!isAuthorized) return {error: 'Unauthorized', status: 401, ok: false};
+            if(userRole===ROLES.SCHOOL_ADMIN){
+                const schoolId = __longToken?.schoolId;
+                const students = await this.mongomodels.student.find({schoolId});
+                return {
+                    students
+                };
+            }
             const students = await this.mongomodels.student.find({});
             return {
                 students
